requirements.txt
django==2.1.7
psycopg2==2.8.4
Pillow==6.2.1
djangorestframework==3.11.0
requests==2.7.0